+++ src/main/java/ladysnake/blast/common/entity/StripminerEntity.java
@@ -3,6 +3,7 @@ package ladysnake.blast.common.entity;
 import ladysnake.blast.common.Blast;
 import ladysnake.blast.common.block.StripminerBlock;
 import ladysnake.blast.common.init.BlastBlocks;
+import ladysnake.blast.common.init.BlastEntities;
 import ladysnake.blast.common.init.BlastItems;
 import ladysnake.blast.common.world.CustomExplosion;
 import net.minecraft.block.BlockState;
@@ -78,7 +79,7 @@ public class StripminerEntity extends BombEntity {
 
     @Override
     protected Item getDefaultItem() {
-        return BlastItems.BOMB;
+        return this.getType() == BlastEntities.COLD_DIGGER ? BlastBlocks.COLD_DIGGER.asItem() : BlastBlocks.STRIPMINER.asItem();
     }
 
     @Override