+++ src/main/java/ladysnake/blast/common/world/CustomExplosion.java
@@ -36,7 +36,7 @@ import net.minecraft.world.explosion.Explosion;
 import java.util.*;
 
 public class CustomExplosion extends Explosion {
-    public final Explosion.DestructionType blockDestructionType;
+    public final Explosion.DestructionType destructionType;
     public final Random random;
     public final World world;
     public final double x;
@@ -61,7 +61,7 @@ public class CustomExplosion extends Explosion {
         this.x = x;
         this.y = y;
         this.z = z;
-        this.blockDestructionType = destructionType;
+        this.destructionType = destructionType;
         this.damageSource = DamageSource.explosion(this);
     }
 
@@ -188,7 +188,7 @@ public class CustomExplosion extends Explosion {
 
     public void affectWorld(boolean boolean_1) {
         this.world.playSound(null, this.x, this.y, this.z, SoundEvents.ENTITY_GENERIC_EXPLODE, SoundCategory.BLOCKS, 4.0F, (1.0F + (this.world.random.nextFloat() - this.world.random.nextFloat()) * 0.2F) * 0.7F);
-        boolean boolean_2 = this.blockDestructionType != DestructionType.NONE;
+        boolean boolean_2 = this.destructionType != DestructionType.NONE;
         if (this.power >= 2.0F && boolean_2) {
             this.world.addParticle(ParticleTypes.EXPLOSION_EMITTER, this.x, this.y, this.z, 1.0D, 0.0D, 0.0D);
         } else {
@@ -242,7 +242,7 @@ public class CustomExplosion extends Explosion {
                         }
 
                         LootContext.Builder builder = (new LootContext.Builder((ServerWorld) this.world)).random(this.world.random).parameter(LootContextParameters.ORIGIN, Vec3d.of(blockPos)).parameter(LootContextParameters.TOOL, itemStack).optionalParameter(LootContextParameters.BLOCK_ENTITY, blockEntity).optionalParameter(LootContextParameters.THIS_ENTITY, this.entity);
-                        if (this.blockDestructionType == DestructionType.DESTROY) {
+                        if (this.destructionType == DestructionType.DESTROY) {
                             builder.parameter(LootContextParameters.EXPLOSION_RADIUS, this.power);
                         }